requirements.txt
llama-index
openai
langchain
python-dotenv